so  many changes. Refactored packages, externalized lib, minor DemoApp Enhancements wip Maven publish
