+++ .idea/gradle.xml
@@ -5,11 +5,12 @@
       <GradleProjectSettings>
         <option name="distributionType" value="DEFAULT_WRAPPED" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
-        <option name="gradleJvm" value="1.7" />
+        <option name="gradleJvm" value="1.8" />
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
-            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/tagViewDemoApp" />
+            <option value="$PROJECT_DIR$/tagviewlib" />
           </set>
         </option>
       </GradleProjectSettings>
