--- app/src/main/java/cuneyt/example/com/tagview/Tag/OnTagClickListener.java
@@ -1,8 +1,0 @@
-package cuneyt.example.com.tagview.Tag;
-
-/**
- * listener for tag delete
- */
-public interface OnTagClickListener {
-	void onTagClick(Tag tag, int position);
-}
