--- app/src/main/java/cuneyt/example/com/tagview/Tag/OnTagDeleteListener.java
@@ -1,8 +1,0 @@
-package cuneyt.example.com.tagview.Tag;
-
-/**
- * listener for tag delete
- */
-public interface OnTagDeleteListener {
-	void onTagDeleted(TagView view, Tag tag, int position);
-}
