--- app/src/main/java/cuneyt/example/com/tagview/Tag/Utils.java
@@ -1,14 +1,0 @@
-package cuneyt.example.com.tagview.Tag;
-
-import android.content.Context;
-import android.util.DisplayMetrics;
-import android.util.TypedValue;
-
-public class Utils {
-
-	public static int dipToPx(Context c,float dipValue) {
-		DisplayMetrics metrics = c.getResources().getDisplayMetrics();
-		return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dipValue, metrics);
-	}
-
-}
