--- app/src/main/res/layout/activity_main.xml
@@ -1,34 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
-    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity">
-
-    <EditText
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/editText"
-        android:layout_alignRight="@+id/tags_laoyut"
-        android:layout_alignEnd="@+id/tags_laoyut"
-        android:layout_alignLeft="@+id/tags_laoyut"
-        android:layout_alignStart="@+id/tags_laoyut" />
-
-    <ScrollView
-        android:id="@+id/tags_laoyut"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginLeft="6dp"
-        android:background="@android:color/white"
-        android:visibility="visible"
-        android:layout_below="@+id/editText">
-
-        <cuneyt.example.com.tagview.Tag.TagView
-            android:id="@+id/tag_group"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_margin="10dp" />
-
-    </ScrollView>
-
-</RelativeLayout>
