--- app/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<resources>
-    <string name="app_name">TagView</string>
-
-    <string name="hello_world">Hello world!</string>
-    <string name="action_settings">Settings</string>
-
-</resources>
