+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app'
+include ':tagViewDemoApp', ':tagviewlib'
