+++ tagViewDemoApp/build.gradle
@@ -1,0 +1,40 @@
+apply plugin: 'com.android.application'
+
+// ext is a gradle closure allowing the declaration of global properties
+ext {
+    PUBLISH_GROUP_ID = 'it.angelic'
+    PUBLISH_ARTIFACT_ID = 'my-library-example'
+    PUBLISH_VERSION = '1.0.0'
+}
+
+
+apply from: 'https://raw.githubusercontent.com/blundell/release-android-library/master/android-release-aar.gradle'
+//facciamo cosi
+//http://blog.blundellapps.co.uk/locally-release-an-android-library-for-jcenter-or-maven-central-inclusion/
+
+
+android {
+    compileSdkVersion 23
+    buildToolsVersion "23.0.2"
+
+    defaultConfig {
+        minSdkVersion 15
+        targetSdkVersion 23
+        versionCode 1
+        versionName project.PUBLISH_VERSION
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+dependencies {
+    compile fileTree(include: ['*.jar'], dir: 'libs')
+    compile 'com.android.support:appcompat-v7:23.2.0'
+    compile 'com.jakewharton:butterknife:6.1.0'
+    compile 'me.drakeet.materialdialog:library:1.2.2'
+    compile project(':tagviewlib')
+}
