+++ tagViewDemoApp/proguard-rules.pro
