+++ tagViewDemoApp/src/main/res/values-w820dp/tagview_attr.xml
