+++ tagViewDemoApp/src/main/res/values/dimens.xml
