+++ tagViewDemoApp/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">TagViewDemo</string>
+    <string name="action_settings">Settings</string>
+</resources>
