+++ tagViewDemoApp/src/main/res/values/styles.xml
