+++ tagviewlib/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="it.angelic.tagviewlib">
+
+
+</manifest>
