+++ tagviewlib/src/main/java/Tag/OnTagClickListener.java
@@ -1,0 +1,8 @@
+package Tag;
+
+/**
+ * listener for tag delete
+ */
+public interface OnTagClickListener {
+	void onTagClick(Tag tag, int position);
+}
