+++ tagviewlib/src/main/java/Tag/OnTagDeleteListener.java
@@ -1,0 +1,8 @@
+package Tag;
+
+/**
+ * listener for tag delete
+ */
+public interface OnTagDeleteListener {
+	void onTagDeleted(TagView view, Tag tag, int position);
+}
