+++ tagviewlib/src/main/res/layout/tagview_item.xml
